requirements.txt
Django>=1.8
GitPython>=1.0
